Fix build and add CI
