+++ .github/workflows/dotnet.yml
@@ -1,0 +1,17 @@
+name: .NET Build
+
+on:
+  push:
+    branches: [ "main", "master", "work" ]
+  pull_request:
+
+jobs:
+  build:
+    runs-on: windows-latest
+    steps:
+    - uses: actions/checkout@v3
+    - uses: actions/setup-dotnet@v3
+      with:
+        dotnet-version: '7.0.x'
+    - run: dotnet restore SteamManager/SteamManager.csproj
+    - run: dotnet build SteamManager/SteamManager.csproj --configuration Release --no-restore
