+++ README
@@ -6,8 +6,10 @@ launch Counter-Strike 2 from any account and send trade offers. The interface
 can be switched between English and Russian.
 
 ## Building
-1. Install the .NET 7 SDK with Windows desktop support.
+1. Install the .NET 7 SDK with Windows desktop support. On Linux you can use the [dotnet-install script](https://learn.microsoft.com/dotnet/core/tools/dotnet-install-script).
 2. Run `dotnet restore` and then `dotnet run -p SteamManager/SteamManager.csproj`.
+
+Continuous builds run on GitHub Actions using the included workflow in `.github/workflows/dotnet.yml`.
 
 ## Configuration
 Create an `accounts.json` file with the following structure:
