+++ SteamManager/SteamManager.csproj
@@ -2,6 +2,7 @@
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
     <TargetFramework>net7.0-windows</TargetFramework>
+    <EnableWindowsTargeting>true</EnableWindowsTargeting>
     <UseWPF>true</UseWPF>
     <RootNamespace>SteamManager</RootNamespace>
   </PropertyGroup>
